create templates for ui elements
