+++ frontend/src/components/ui/Button.tsx
@@ -1,10 +1,31 @@
-import React from 'react'
 
 type ButtonProps = React.ButtonHTMLAttributes<HTMLButtonElement> & {
-  vaiant: 'greean'|'blue'|'gray';
-  size: 'sm'|'md'|'lg';
-}
+  variant?: "green" | "blue" | "gray";
+  size?: "sm" | "md" | "lg";
+};
 
 const variantClasses = {
-  green
-}
+  green: "bg-green-500 hover:bg-green-600 text-white",
+  blue: "bg-blue-500 hover:bg-blue-600 text-white",
+  gray: "bg-gray-300 hover:bg-gray-400 text-gray-800",
+};
+
+const sizeClasses = {
+  sm: "px-3 py-1 text-sm",
+  md: "px-4 py-2 text-base",
+  lg: "px-6 py-3 text-lg",
+};
+
+export const Button: React.FC<ButtonProps> = ({
+  variant = "green",
+  size = "md",
+  className = "",
+  ...props
+}) => {
+  return (
+    <button
+      className={`rounded-md font-medium transition-colors ${variantClasses[variant]} ${sizeClasses[size]} ${className}`}
+      {...props}
+    />
+  );
+};
